Move brand-context-tools out of the framework
Pure move, no behaviour change: `src/lib/agent/tools/brand-context-tools.ts`
becomes `src/lib/server/brand-context-tools.ts`, and its five importers
follow.

It never belonged to the framework. It imports `ai`, `zod` and a
Supabase type, and nothing else — but it lives inside `src/lib/agent`,
which is on the list to delete, and four things that must SURVIVE depend
on it: Motion Video, the Media Generator, the UGC planner and the UGC
agent. Those are content generation, which is the product. Deleting
`src/lib/agent` with this file inside it would have taken the four reads
those generators use to know what the brand is.

The test moves with it and its relative paths are re-pointed; the two
assertions that read files still inside `src/lib/agent/tools` now say so
explicitly.
